+++ src/ExerciseComponents/TodoApp/TodoList/TodoList.tsx
@@ -2,6 +2,7 @@ import { useMemo } from "react";
 import { Button, Checkbox, HStack, List, ListItem } from "@chakra-ui/react";
 import { type TodoFilterStatus } from "../TodoListFilter/TodoListFilter.type";
 import { useTodoList } from "./useTodoList";
+import { filterByQuery, filterByStatus } from "./FilterTodoList";
 
 type TodoListProps = {
   query: string | undefined;
@@ -9,7 +10,7 @@ type TodoListProps = {
 };
 
 export function TodoList({ query, status }: TodoListProps) {
-  const { todoList, toggleTodo, deleteTodo, filterByStatus, filterByQuery } = useTodoList();
+  const { todoList, toggleTodo, deleteTodo } = useTodoList();
 
   const filteredTodoList = useMemo(() => {
     const filteredByStatusList = filterByStatus(todoList, status);
